requirements.txt
fastapi
uvicorn
sqlalchemy
pymysql
python-dotenv
pydantic
pydantic[email]
pydantic-settings
cryptography
aiomysql
passlib
email-validator
mysqlclient
bcrypt
alembic
greenlet
python-jose[cryptography]
python-multipart